split for pbswitch routines
	modified:   hoosg.cpp
	modified:   main.cpp
	new file:   pbswitch.cpp
	new file:   pbswitch.h

On branch dvlp-aa
