+++ pbswitch-a.d/examples/uno/src/main.cpp
@@ -1,6 +1,7 @@
 /* main.cpp - main program 9 September 2024 */
-#include "hoosg.h"
 #include <Arduino.h>
+#include "hoosg.h"
+#include "pbswitch.h"
 
 void setup() {
   initSerial();
@@ -9,56 +10,8 @@ void setup() {
   Serial.println(" blink() already ran.");
 }
 
-int counter = -1;
-
-void actuate() {
-  counter++;
-  newLine();
-  Serial.print(counter);
-  Serial.print("     TADA");
-}
-
-bool readingSwitch() {
-  bool held = false;
-  held = scanPbSwitch();
-  return held;
-}
-
-int hbeat = -1;
-int iHBeat = -1;
-
-void showHBeat() {
-  Serial.write(' ');
-  Serial.write('.');
-  Serial.write(' ');
-}
-
-void iHeartbeat() {
-  int hbInterval = (128 - 10); /* 16384 */
-  hbeat++;
-  if (hbeat > hbInterval) {
-    showHBeat();
-    hbeat = -1;
-  }
-}
-
-void heartbeat() {
-  int iHbIval = (4096 - 10);
-  iHBeat++;
-  if (iHBeat > iHbIval) {
-    iHeartbeat();
-    iHBeat = -1;
-  }
-}
-
 void loop() {
-  bool switchActive = false;
-  switchActive = readingSwitch();
-  if (switchActive == true) {
-    actuate();
-    switchActive = false;
-  }
-  heartbeat();
+  pbSwitchLoop();
 }
 
 /* reformat ctrl shift i as in india which F1 menu tells you */
